[web] discover cloud iac integration hook
change name to useEnrollCloudIntegration, fix linter error

Fix IaCIntegrationOverview story broken by react-router update
