+++ web/packages/teleport/src/Discover/Overview/IaCIntegrationOverview.story.tsx
@@ -17,16 +17,13 @@
  */
 
 import { http, HttpResponse } from 'msw';
-import { MemoryRouter, Route } from 'react-router';
+import { Routes, Route } from 'react-router';
 
-import { InfoGuidePanelProvider } from 'shared/components/SlidingSidePanel/InfoGuide';
-import {
-  ToastNotificationProvider,
-  ToastNotifications,
-} from 'shared/components/ToastNotification';
+import { ToastNotifications } from 'shared/components/ToastNotification';
 
 import cfg from 'teleport/config';
 import { ContentMinWidth } from 'teleport/Main/Main';
+import { TeleportProviderBasic } from 'teleport/mocks/providers';
 import { IntegrationKind } from 'teleport/services/integrations';
 
 import { IaCIntegrationOverview } from './IaCIntegrationOverview';
@@ -43,18 +40,14 @@ const path = cfg.routes.integrationOverview;
 
 function Component() {
   return (
-    <MemoryRouter initialEntries={initialEntries}>
-      <ToastNotificationProvider>
-        <ToastNotifications />
-        <InfoGuidePanelProvider>
-          <ContentMinWidth>
-            <Route path={path}>
-              <IaCIntegrationOverview />
-            </Route>
-          </ContentMinWidth>
-        </InfoGuidePanelProvider>
-      </ToastNotificationProvider>
-    </MemoryRouter>
+    <TeleportProviderBasic initialEntries={initialEntries}>
+      <ToastNotifications />
+      <ContentMinWidth>
+        <Routes>
+          <Route path={path} element={<IaCIntegrationOverview />} />
+        </Routes>
+      </ContentMinWidth>
+    </TeleportProviderBasic>
   );
 }
